fix: split formatter from admin cli
