+++ crates/reinhardt-admin-cli/Cargo.toml
@@ -25,14 +25,6 @@ prettyplease = "0.2"
 colored = { workspace = true }
 ctrlc = "3.4"
 regex = "1.10"
-tree-sitter = "0.26"
-tree-sitter-rust = "0.24.2"
-tree-sitter-reinhardt-page = { workspace = true }
-tree-sitter-reinhardt-form = { workspace = true }
-tree-sitter-reinhardt-head = { workspace = true }
-topiary-core = { workspace = true }
-topiary-tree-sitter-facade = { workspace = true }
-zeroize = "1.8"
 anyhow = { workspace = true }
 
 [dev-dependencies]
