+++ crates/reinhardt-admin-cli/README.md
@@ -86,7 +86,7 @@ reinhardt-admin plugin info auth-delion --remote
 
 # Install a plugin
 reinhardt-admin plugin install auth-delion
-reinhardt-admin plugin install auth-delion --version 0.1.0-rc.30
+reinhardt-admin plugin install auth-delion --version 0.1.3
 
 # Remove a plugin
 reinhardt-admin plugin remove auth-delion
@@ -105,37 +105,39 @@ reinhardt-admin plugin update --all
 
 ### Format All Code
 
-Format all Rust files in the project (Rust + `page!` DSL):
+Use `reinhardt-formatter` to format all Rust files in the project. Reinhardt
+DSL macros are formatted with tree-sitter grammars and Topiary queries before
+`cargo fmt --all` formats the surrounding Rust code:
 
 ```bash
 # Format all files in the project
-reinhardt-admin fmt-all
+reinhardt-formatter fmt-all
 
 # Check formatting without modifying files
-reinhardt-admin fmt-all --check
+reinhardt-formatter fmt-all --check
 ```
 
-### Format page! Macro DSL
+### Format Reinhardt Macro DSLs
 
-Format `page!` macro DSL in your source files:
+Format `page!`, `form!`, and `head!` macro DSLs in your source files:
 
 ```bash
 # Format all Rust files in the current directory
-reinhardt-admin fmt .
+reinhardt-formatter fmt .
 
 # Format a specific file
-reinhardt-admin fmt src/main.rs
+reinhardt-formatter fmt src/main.rs
 
 # Check formatting without modifying files
-reinhardt-admin fmt --check .
+reinhardt-formatter fmt --check .
 
 # Show all files (including unchanged)
-reinhardt-admin fmt -v .
+reinhardt-formatter fmt -v .
 ```
 
 #### Ignore Markers
 
-You can control which `page!` macros should be skipped during formatting by using special comment markers:
+You can control which Reinhardt DSL macros should be skipped during formatting by using special comment markers:
 
 ##### File-Wide Ignore
 
@@ -151,7 +153,7 @@ page!(|| { div{bad_format} })  // Will not be formatted
 
 ##### Range Ignore
 
-Skip formatting for multiple macros within a range using `// reinhardt-fmt: off` and `// reinhardt-fmt: on`:
+Skip formatting for multiple DSL macros within a range using `// reinhardt-fmt: off` and `// reinhardt-fmt: on`:
 
 ```rust
 // reinhardt-fmt: off
@@ -164,7 +166,7 @@ page!(|| { p { good } })  // This will be formatted
 
 ##### Individual Macro Ignore
 
-Skip formatting for a specific macro by adding `// reinhardt-fmt: ignore` on the line immediately before it:
+Skip formatting for a specific DSL macro by adding `// reinhardt-fmt: ignore` on the line immediately before it:
 
 ```rust
 // reinhardt-fmt: ignore
@@ -185,7 +187,7 @@ When multiple markers are present, they are applied in this priority order:
 
 - Markers are case-sensitive and spaces are optional (e.g., `//reinhardt-fmt:ignore` also works)
 - Individual markers must be on the line **immediately before** the macro (no blank lines)
-- Range markers can span multiple macros
+- Range markers can span multiple DSL macros
 - Nested `off` markers will generate a warning but use the first `off` position
 - Unclosed ranges (missing `on`) will extend to the end of the file
 
@@ -207,7 +209,7 @@ When multiple markers are present, they are applied in this priority order:
 #### Example Output
 
 ```bash
-$ reinhardt-admin fmt .
+$ reinhardt-formatter fmt .
 [1/47] Formatted: src/main.rs
 [2/47] Formatted: src/config/settings.rs
 [3/47] Error src/broken.rs: Parse error
