+++ crates/reinhardt-admin-cli/src/main.rs
@@ -24,17 +24,13 @@
 //! reinhardt-admin startproject myproject --with-rest
 //! reinhardt-admin startproject myproject --with-pages
 //! reinhardt-admin startapp myapp --with-rest
-//! reinhardt-admin fmt src/
+//! reinhardt-formatter fmt src/
 //! reinhardt-admin --help
 //! ```
 
-mod format_engine;
-mod formatter;
-mod utils;
-
 use reinhardt_admin_cli::migrate_v2;
 
-use std::path::{Path, PathBuf};
+use std::path::PathBuf;
 
 use clap::{Parser, Subcommand, ValueEnum};
 use colored::Colorize;
@@ -44,7 +40,6 @@ use reinhardt_commands::{
 	PluginSearchCommand, PluginUpdateCommand, StartAppCommand, StartProjectCommand,
 };
 use std::process;
-use zeroize::Zeroize;
 
 /// Project/app architecture type used with `--template`.
 #[derive(Clone, Debug, ValueEnum)]
@@ -681,259 +676,19 @@ fn run_fmt(
 	backup: bool,
 	verbosity: u8,
 ) -> CommandResult<()> {
-	use format_engine::{FormatEngine, RustfmtOptions};
-	use formatter::collect_rust_files;
-
-	// Validate config path if provided
-	if let Some(ref cp) = config_path {
-		validate_config_path(cp).map_err(|e| {
-			reinhardt_commands::CommandError::ExecutionError(format!("Invalid config path: {}", e))
-		})?;
-		check_file_size(cp, MAX_CONFIG_FILE_SIZE)
-			.map_err(reinhardt_commands::CommandError::ExecutionError)?;
+	let mut command = formatter_delegate_command(verbosity);
+	command.arg("fmt").arg(path);
+	push_flag(&mut command, "--check", check);
+	if !with_rustfmt {
+		command.arg("--with-rustfmt=false");
 	}
-
-	let files = collect_rust_files(&path).map_err(|e| {
-		reinhardt_commands::CommandError::ExecutionError(format!(
-			"Failed to collect files in {}: {}",
-			display_path(&path),
-			sanitize_error(&e)
-		))
-	})?;
-
-	if files.is_empty() {
-		if verbosity > 0 {
-			println!("No Rust files found in {}", display_path(&path));
-		}
-		return Ok(());
-	}
-
-	// Resolve config path
-	let resolved_config_path = config_path.or_else(|| find_rustfmt_config(&path));
-
-	let options = RustfmtOptions {
-		config_path: resolved_config_path.clone(),
-		edition,
-		style_edition,
-		config,
-		color: Some(color),
-	};
-
-	if verbosity > 0
-		&& let Some(ref p) = resolved_config_path
-	{
-		println!("Using rustfmt config: {}", display_path(p));
-	}
-
-	let formatter = FormatEngine::new();
-	let mut formatted_count = 0;
-	let mut unchanged_count = 0;
-	let mut ignored_count = 0;
-	let mut error_count = 0;
-
-	let total_files = files.len();
-
-	for (index, file_path) in files.iter().enumerate() {
-		let progress = format!("[{}/{}]", index + 1, total_files);
-
-		// Check file size before reading to prevent OOM
-		if let Err(e) = check_file_size(file_path, MAX_SOURCE_FILE_SIZE) {
-			eprintln!(
-				"{} {} {}: {}",
-				progress.bright_blue(),
-				"Skipped:".yellow(),
-				display_path(file_path),
-				e
-			);
-			error_count += 1;
-			continue;
-		}
-
-		let original_content = std::fs::read_to_string(file_path).map_err(|e| {
-			reinhardt_commands::CommandError::ExecutionError(format!(
-				"Failed to read {}: {}",
-				mask_path(file_path),
-				sanitize_error(&e.to_string())
-			))
-		})?;
-
-		// Check for file-wide ignore marker BEFORE any processing.
-		// When with_rustfmt is enabled, skip only the DSL pass but still
-		// run rustfmt below so plain Rust formatting is applied.
-		let ignore_all = formatter.has_ignore_all_marker(&original_content);
-		if ignore_all && !with_rustfmt {
-			ignored_count += 1;
-			if verbosity > 0 {
-				println!(
-					"{} {} {} (reinhardt-fmt: ignore-all)",
-					progress.bright_blue(),
-					"Ignored:".yellow(),
-					display_path(file_path)
-				);
-			}
-			continue;
-		}
-
-		let dsl_result = if ignore_all {
-			// ignore-all with rustfmt: skip DSL pass, feed original to rustfmt
-			original_content.clone()
-		} else {
-			match formatter.format(&original_content) {
-				Ok(result) => {
-					if let Some(reason) = &result.skipped {
-						if !with_rustfmt {
-							ignored_count += 1;
-							println!(
-								"{} {} {} ({})",
-								progress.bright_blue(),
-								"Ignored:".yellow(),
-								display_path(file_path),
-								reason
-							);
-							continue;
-						}
-						original_content.clone()
-					} else {
-						result.content
-					}
-				}
-				Err(e) => {
-					eprintln!(
-						"{} {} {}: {}",
-						progress.bright_blue(),
-						"Error".red(),
-						display_path(file_path),
-						sanitize_error(&e.to_string())
-					);
-					error_count += 1;
-					continue;
-				}
-			}
-		};
-
-		let final_result = if with_rustfmt {
-			match run_rustfmt(&dsl_result, &options) {
-				Ok(output) => output,
-				Err(e) => {
-					eprintln!(
-						"{} {} {}: rustfmt failed: {}",
-						progress.bright_blue(),
-						"Error".red(),
-						display_path(file_path),
-						sanitize_error(&e)
-					);
-					error_count += 1;
-					continue;
-				}
-			}
-		} else {
-			dsl_result
-		};
-
-		// Compare with original
-		if final_result != original_content {
-			if check {
-				// Check mode: report unformatted files
-				println!("{} Would format: {}", progress, display_path(file_path));
-				formatted_count += 1;
-			} else {
-				// Backup if requested (with RAII guard to clean up on failure)
-				let mut _backup_guard = None;
-				if backup {
-					let backup_path = create_temp_backup_path(file_path);
-					create_secure_backup(file_path, &backup_path).map_err(|e| {
-						reinhardt_commands::CommandError::ExecutionError(format!(
-							"Failed to backup {}: {}",
-							mask_path(file_path),
-							e
-						))
-					})?;
-					_backup_guard = Some(utils::BackupGuard::new(backup_path));
-				}
-
-				// Format mode: write changes atomically (write to temp, then rename)
-				utils::atomic_write(file_path, &final_result).map_err(|e| {
-					reinhardt_commands::CommandError::ExecutionError(format!(
-						"Failed to write {}: {}",
-						mask_path(file_path),
-						sanitize_error(&e.to_string())
-					))
-				})?;
-
-				// Commit the backup guard so the backup is preserved on success
-				if let Some(ref mut guard) = _backup_guard {
-					guard.commit();
-				}
-				// Color output: success in green
-				println!(
-					"{} {} {}",
-					progress.bright_blue(),
-					"Formatted:".green(),
-					display_path(file_path)
-				);
-				formatted_count += 1;
-			}
-		} else {
-			unchanged_count += 1;
-			if verbosity > 0 {
-				println!(
-					"{} {} {}",
-					progress.bright_blue(),
-					"Unchanged:".dimmed(),
-					display_path(file_path)
-				);
-			}
-		}
-	}
-
-	// Always show summary (remove verbosity condition)
-	println!();
-	if check {
-		println!(
-			"{}: {} would be formatted, {} unchanged, {} ignored, {} errors",
-			"Summary".bright_cyan(),
-			formatted_count.to_string().yellow(),
-			unchanged_count,
-			ignored_count,
-			if error_count > 0 {
-				error_count.to_string().red()
-			} else {
-				error_count.to_string().green()
-			}
-		);
-	} else {
-		println!(
-			"{}: {} formatted, {} unchanged, {} ignored, {} errors",
-			"Summary".bright_cyan(),
-			if formatted_count > 0 {
-				formatted_count.to_string().green()
-			} else {
-				formatted_count.to_string().dimmed()
-			},
-			unchanged_count,
-			ignored_count,
-			if error_count > 0 {
-				error_count.to_string().red()
-			} else {
-				error_count.to_string().dimmed()
-			}
-		);
-	}
-
-	if check && formatted_count > 0 {
-		return Err(reinhardt_commands::CommandError::ExecutionError(
-			"Some files are not properly formatted".to_string(),
-		));
-	}
-
-	if error_count > 0 {
-		return Err(reinhardt_commands::CommandError::ExecutionError(format!(
-			"{} files had formatting errors",
-			error_count
-		)));
-	}
-
-	Ok(())
+	push_optional_path(&mut command, "--config-path", config_path.as_ref());
+	push_optional_str(&mut command, "--edition", edition.as_deref());
+	push_optional_str(&mut command, "--style-edition", style_edition.as_deref());
+	push_optional_str(&mut command, "--config", config.as_deref());
+	push_optional_str(&mut command, "--color", Some(&color));
+	push_flag(&mut command, "--backup", backup);
+	run_formatter_delegate(command)
 }
 
 /// Format all code: DSL macros via Topiary, then Rust via `cargo fmt --all`.
@@ -948,706 +703,57 @@ fn run_fmt_all(
 	backup: bool,
 	verbosity: u8,
 ) -> CommandResult<()> {
-	use format_engine::FormatEngine;
-	use formatter::{collect_rust_files, nested_workspace_roots};
-	use std::collections::HashMap;
-	use std::process::{Command, Stdio};
+	let mut command = formatter_delegate_command(verbosity);
+	command.arg("fmt-all");
+	push_flag(&mut command, "--check", check);
+	push_optional_path(&mut command, "--config-path", config_path.as_ref());
+	push_optional_str(&mut command, "--edition", edition.as_deref());
+	push_optional_str(&mut command, "--style-edition", style_edition.as_deref());
+	push_optional_str(&mut command, "--config", config.as_deref());
+	push_optional_str(&mut command, "--color", Some(&color));
+	push_flag(&mut command, "--backup", backup);
+	run_formatter_delegate(command)
+}
 
-	// Validate config path if provided
-	if let Some(ref cp) = config_path {
-		validate_config_path(cp).map_err(|e| {
-			reinhardt_commands::CommandError::ExecutionError(format!("Invalid config path: {}", e))
-		})?;
-		check_file_size(cp, MAX_CONFIG_FILE_SIZE)
-			.map_err(reinhardt_commands::CommandError::ExecutionError)?;
+fn formatter_delegate_command(verbosity: u8) -> process::Command {
+	let mut command = process::Command::new("reinhardt-formatter");
+	for _ in 0..verbosity {
+		command.arg("-v");
 	}
+	command
+}
 
-	// Find project root
-	let project_root = find_project_root().ok_or_else(|| {
-		reinhardt_commands::CommandError::ExecutionError(
-			"Could not find project root (no Cargo.toml found)".to_string(),
-		)
-	})?;
-
-	if verbosity > 0 {
-		println!("Project root: {}", display_path(&project_root));
+fn push_flag(command: &mut process::Command, flag: &str, enabled: bool) {
+	if enabled {
+		command.arg(flag);
 	}
+}
 
-	let files = collect_rust_files(&project_root).map_err(|e| {
+fn push_optional_path(command: &mut process::Command, flag: &str, value: Option<&PathBuf>) {
+	if let Some(value) = value {
+		command.arg(flag).arg(value);
+	}
+}
+
+fn push_optional_str(command: &mut process::Command, flag: &str, value: Option<&str>) {
+	if let Some(value) = value {
+		command.arg(flag).arg(value);
+	}
+}
+
+fn run_formatter_delegate(mut command: process::Command) -> CommandResult<()> {
+	let status = command.status().map_err(|err| {
 		reinhardt_commands::CommandError::ExecutionError(format!(
-			"Failed to collect files: {}",
-			sanitize_error(&e)
+			"failed to run reinhardt-formatter: {err}. Install or build the formatter binary to use fmt commands."
 		))
 	})?;
 
-	// Exclude files that belong to a separate (nested) cargo workspace such as
-	// `examples/` or the trybuild fixture crates. The `cargo fmt --all` pass
-	// below only formats the root workspace, so applying just the DSL pass to
-	// these files would leave them rustfmt-unformatted and contradict their own
-	// `fmt`/`cargo fmt` tasks (e.g. `cargo make fmt-check-examples`). Each
-	// sub-workspace is formatted by its dedicated task instead. Refs #5051.
-	let nested_roots = nested_workspace_roots(&project_root);
-	let files: Vec<PathBuf> = files
-		.into_iter()
-		.filter(|file| !nested_roots.iter().any(|root| file.starts_with(root)))
-		.collect();
-
-	if files.is_empty() {
-		if verbosity > 0 {
-			println!("No Rust files found in {}", display_path(&project_root));
-		}
-		return Ok(());
-	}
-
-	let formatter = FormatEngine::new();
-
-	// Acquire a lock file to prevent concurrent format operations (TOCTOU mitigation)
-	let lock_path = project_root.join(".reinhardt-fmt.lock");
-	let _lock_file = acquire_format_lock(&lock_path).map_err(|e| {
-		reinhardt_commands::CommandError::ExecutionError(format!(
-			"Failed to acquire format lock: {}. Another format operation may be in progress.",
-			e
-		))
-	})?;
-
-	// Store original contents for comparison and rollback
-	let mut original_contents: HashMap<PathBuf, String> = HashMap::new();
-
-	let total_files = files.len();
-	let mut dsl_file_count = 0;
-	let mut error_count = 0;
-
-	if verbosity > 0 {
-		println!(
-			"{} Phase 1: Formatting Reinhardt DSL macros...",
-			"[Step 1/2]".bright_blue()
-		);
-	}
-
-	for file_path in &files {
-		if let Err(e) = check_file_size(file_path, MAX_SOURCE_FILE_SIZE) {
-			eprintln!("{} Skipping oversized file: {}", "Warning:".yellow(), e);
-			error_count += 1;
-			continue;
-		}
-
-		let original_content = std::fs::read_to_string(file_path).map_err(|e| {
-			reinhardt_commands::CommandError::ExecutionError(format!(
-				"Failed to read {}: {}",
-				mask_path(file_path),
-				sanitize_error(&e.to_string())
-			))
-		})?;
-
-		// Store original content for comparison
-		original_contents.insert(file_path.clone(), original_content.clone());
-
-		let format_result = match formatter.format(&original_content) {
-			Ok(result) => result,
-			Err(e) => {
-				eprintln!(
-					"{} DSL format failed for {}: {}",
-					"Error:".red(),
-					display_path(file_path),
-					sanitize_error(&e.to_string())
-				);
-				error_count += 1;
-				continue;
-			}
-		};
-
-		if !format_result.contains_dsl_macro || format_result.skipped.is_some() {
-			continue;
-		}
-
-		if format_result.content != original_content {
-			utils::atomic_write(file_path, &format_result.content).map_err(|e| {
-				reinhardt_commands::CommandError::ExecutionError(format!(
-					"Failed to write DSL-formatted content to {}: {}",
-					mask_path(file_path),
-					sanitize_error(&e.to_string())
-				))
-			})?;
-			dsl_file_count += 1;
-		}
-	}
-
-	if verbosity > 0 {
-		println!(
-			"{} Phase 2: Running cargo fmt --all...",
-			"[Step 2/2]".bright_blue()
-		);
-	}
-
-	let mut cmd = Command::new("cargo");
-	cmd.arg("fmt").arg("--all");
-	cmd.current_dir(&project_root);
-	cmd.stdout(Stdio::inherit());
-	cmd.stderr(Stdio::inherit());
-
-	// Add rustfmt options via "--" separator
-	let has_rustfmt_options = config_path.is_some()
-		|| edition.is_some()
-		|| style_edition.is_some()
-		|| config.is_some()
-		|| color != "auto";
-
-	if has_rustfmt_options {
-		cmd.arg("--");
-
-		if let Some(ref path) = config_path {
-			cmd.arg("--config-path").arg(path);
-		}
-
-		if let Some(ref ed) = edition {
-			cmd.arg("--edition").arg(ed);
-		}
-
-		if let Some(ref se) = style_edition {
-			cmd.arg("--style-edition").arg(se);
-		}
-
-		if let Some(ref cfg) = config {
-			cmd.arg("--config").arg(cfg);
-		}
-
-		if color != "auto" {
-			cmd.arg("--color").arg(&color);
-		}
-	}
-
-	if verbosity > 1 {
-		// Only show command details at high verbosity to avoid leaking config paths
-		println!("  Running: cargo fmt --all");
-	}
-
-	let cargo_fmt_result = cmd.output();
-
-	let output = match cargo_fmt_result {
-		Ok(output) => output,
-		Err(e) => {
-			eprintln!(
-				"{} cargo fmt failed: {}",
-				"Error:".red(),
-				sanitize_error(&e.to_string())
-			);
-			let rollback_paths: Vec<PathBuf> = original_contents.keys().cloned().collect();
-			let rollback_errors = utils::rollback_files(&rollback_paths, &original_contents);
-			utils::report_rollback_errors(&rollback_errors);
-			return Err(reinhardt_commands::CommandError::ExecutionError(
-				"cargo fmt failed to execute".to_string(),
-			));
-		}
-	};
-
-	if !output.status.success() {
-		let stderr = String::from_utf8_lossy(&output.stderr);
-		let sanitized_stderr = sanitize_error(&stderr);
-		eprintln!(
-			"{} cargo fmt exited with error: {}",
-			"Error:".red(),
-			sanitized_stderr
-		);
-		let rollback_paths: Vec<PathBuf> = original_contents.keys().cloned().collect();
-		let rollback_errors = utils::rollback_files(&rollback_paths, &original_contents);
-		utils::report_rollback_errors(&rollback_errors);
-		return Err(reinhardt_commands::CommandError::ExecutionError(format!(
-			"cargo fmt exited with error: {}",
-			sanitized_stderr
-		)));
-	}
-
-	// Compare and count changes
-	let mut formatted_count = 0;
-	let mut unchanged_count = 0;
-
-	for (index, file_path) in files.iter().enumerate() {
-		let progress = format!("[{}/{}]", index + 1, total_files);
-
-		let original_content = match original_contents.get(file_path) {
-			Some(content) => content,
-			None => continue,
-		};
-
-		let current_content = match std::fs::read_to_string(file_path) {
-			Ok(content) => content,
-			Err(_) => continue,
-		};
-
-		if &current_content != original_content {
-			if check {
-				println!("{} Would format: {}", progress, display_path(file_path));
-				// Restore original content in check mode
-				if let Err(e) = std::fs::write(file_path, original_content) {
-					eprintln!(
-						"Warning: failed to restore {} in check mode: {}",
-						file_path.display(),
-						e
-					);
-				}
-			} else if verbosity > 0 {
-				println!(
-					"{} {} {}",
-					progress.bright_blue(),
-					"Formatted:".green(),
-					display_path(file_path)
-				);
-			}
-			formatted_count += 1;
-
-			// Create backup if requested (stored in /tmp with restrictive permissions)
-			// BackupGuard ensures cleanup if the write fails
-			if backup && !check {
-				let backup_path = create_temp_backup_path(file_path);
-				match create_secure_backup(file_path, &backup_path) {
-					Ok(()) => {
-						// Commit immediately since the format write already succeeded
-						let mut guard = utils::BackupGuard::new(backup_path);
-						guard.commit();
-					}
-					Err(e) => {
-						eprintln!(
-							"Warning: failed to create backup for {}: {}",
-							mask_path(file_path),
-							e
-						);
-					}
-				}
-			}
-		} else {
-			unchanged_count += 1;
-			if verbosity > 0 {
-				println!(
-					"{} {} {}",
-					progress.bright_blue(),
-					"Unchanged:".dimmed(),
-					display_path(file_path)
-				);
-			}
-		}
-	}
-
-	// In check mode, restore only files that were modified on disk
-	if check {
-		let all_paths: Vec<PathBuf> = original_contents.keys().cloned().collect();
-		let rollback_errors = utils::rollback_files(&all_paths, &original_contents);
-		utils::report_rollback_errors(&rollback_errors);
-	}
-
-	// Securely clear original contents to prevent sensitive data lingering in memory
-	secure_clear_hashmap(&mut original_contents);
-
-	// Summary
-	println!();
-	if check {
-		println!(
-			"{}: {} would be formatted, {} unchanged, {} DSL files preformatted, {} errors",
-			"Summary".bright_cyan(),
-			formatted_count.to_string().yellow(),
-			unchanged_count,
-			dsl_file_count,
-			if error_count > 0 {
-				error_count.to_string().red()
-			} else {
-				error_count.to_string().green()
-			}
-		);
+	if status.success() {
+		Ok(())
 	} else {
-		println!(
-			"{}: {} formatted, {} unchanged, {} DSL files preformatted, {} errors",
-			"Summary".bright_cyan(),
-			if formatted_count > 0 {
-				formatted_count.to_string().green()
-			} else {
-				formatted_count.to_string().dimmed()
-			},
-			unchanged_count,
-			dsl_file_count,
-			if error_count > 0 {
-				error_count.to_string().red()
-			} else {
-				error_count.to_string().dimmed()
-			}
-		);
-	}
-
-	if check && formatted_count > 0 {
-		return Err(reinhardt_commands::CommandError::ExecutionError(
-			"Some files are not properly formatted".to_string(),
-		));
-	}
-
-	if error_count > 0 {
-		return Err(reinhardt_commands::CommandError::ExecutionError(format!(
-			"{} files had formatting errors",
-			error_count
-		)));
-	}
-
-	Ok(())
-}
-
-/// Maximum number of parent directories to traverse when searching for project root.
-///
-/// Prevents unbounded traversal to the filesystem root, which could
-/// find an unrelated `Cargo.toml` from a different project.
-const MAX_PROJECT_ROOT_DEPTH: usize = 10;
-
-/// Find the project root by searching upward for Cargo.toml.
-///
-/// Uses `std::fs::metadata` instead of `Path::exists()` to avoid TOCTOU
-/// race conditions in the existence check. Traversal is bounded to
-/// `MAX_PROJECT_ROOT_DEPTH` levels to prevent finding unrelated projects.
-fn find_project_root() -> Option<PathBuf> {
-	let current_dir = std::env::current_dir().ok()?;
-	let mut current = current_dir.as_path();
-
-	for _ in 0..MAX_PROJECT_ROOT_DEPTH {
-		if std::fs::metadata(current.join("Cargo.toml")).is_ok() {
-			return Some(current.to_path_buf());
-		}
-		current = current.parent()?;
-	}
-	None
-}
-
-/// Find rustfmt.toml by searching upward from a path.
-///
-/// Uses `std::fs::metadata` instead of `Path::exists()` to avoid TOCTOU
-/// race conditions in the existence check.
-fn find_rustfmt_config(start_path: &Path) -> Option<PathBuf> {
-	let mut current = if start_path.is_file() {
-		start_path.parent()
-	} else {
-		Some(start_path)
-	}?;
-
-	for _ in 0..MAX_PROJECT_ROOT_DEPTH {
-		let config = current.join("rustfmt.toml");
-		if std::fs::metadata(&config).is_ok() {
-			return Some(config);
-		}
-		let hidden_config = current.join(".rustfmt.toml");
-		if std::fs::metadata(&hidden_config).is_ok() {
-			return Some(hidden_config);
-		}
-		current = current.parent()?;
-	}
-	None
-}
-
-/// Run rustfmt on content and return formatted output.
-fn run_rustfmt(content: &str, options: &format_engine::RustfmtOptions) -> Result<String, String> {
-	use std::io::Write;
-	use std::process::{Command, Stdio};
-
-	let mut cmd = Command::new("rustfmt");
-	options.apply_to_command(&mut cmd);
-
-	// Fallback to default edition if no config is specified
-	if options.config_path.is_none() && options.edition.is_none() {
-		cmd.arg("--edition=2024");
-	}
-
-	let mut child = cmd
-		.stdin(Stdio::piped())
-		.stdout(Stdio::piped())
-		.stderr(Stdio::piped())
-		.spawn()
-		.map_err(|e| format!("Failed to spawn rustfmt: {}", e))?;
-
-	if let Some(mut stdin) = child.stdin.take() {
-		stdin
-			.write_all(content.as_bytes())
-			.map_err(|e| format!("Failed to write to rustfmt stdin: {}", e))?;
-	}
-
-	let output = child
-		.wait_with_output()
-		.map_err(|e| format!("Failed to wait for rustfmt: {}", e))?;
-
-	if output.status.success() {
-		String::from_utf8(output.stdout).map_err(|e| format!("Invalid UTF-8 from rustfmt: {}", e))
-	} else {
-		let stderr = String::from_utf8_lossy(&output.stderr);
-		Err(format!("rustfmt failed: {}", stderr))
-	}
-}
-
-/// Create a backup file with restrictive permissions (0600).
-///
-/// # Security
-///
-/// Sets file permissions to 0600 (read/write for owner only) to prevent
-/// unauthorized access to backup files which may contain sensitive code.
-#[cfg(unix)]
-fn create_secure_backup(source: &Path, backup_path: &Path) -> Result<(), std::io::Error> {
-	use std::fs::OpenOptions;
-	use std::io::Read;
-	use std::os::unix::fs::OpenOptionsExt;
-
-	// Read source content
-	let mut content = Vec::new();
-	let mut file = std::fs::File::open(source)?;
-	file.read_to_end(&mut content)?;
-
-	// Create backup with restrictive permissions
-	let mut backup_file = OpenOptions::new()
-		.write(true)
-		.create(true)
-		.truncate(true)
-		.mode(0o600) // Owner read/write only
-		.open(backup_path)?;
-
-	std::io::copy(&mut content.as_slice(), &mut backup_file)?;
-
-	// Zeroize the content buffer to prevent sensitive data lingering in memory
-	content.zeroize();
-
-	Ok(())
-}
-
-/// Create a backup file with restrictive permissions (0600).
-///
-/// # Security
-///
-/// On non-Unix systems, this creates the backup file normally with default permissions.
-/// A warning is logged recommending Unix for production deployments.
-#[cfg(not(unix))]
-fn create_secure_backup(source: &Path, backup_path: &Path) -> Result<(), std::io::Error> {
-	use std::io::Read;
-
-	// Read source content
-	let mut content = Vec::new();
-	let mut file = std::fs::File::open(source)?;
-	file.read_to_end(&mut content)?;
-
-	// Create backup with default permissions (Windows ACLs handle security)
-	std::fs::write(backup_path, &content)?;
-
-	// Zeroize the content buffer
-	content.zeroize();
-
-	Ok(())
-}
-
-/// Create a backup file path in the system temporary directory.
-///
-/// # Security
-///
-/// Stores backup files in `/tmp` instead of the source directory to prevent
-/// backup files from being committed or exposed in the project tree.
-/// Uses a deterministic name based on the source file name so the backup
-/// can be identified if cleanup is interrupted.
-fn create_temp_backup_path(source: &Path) -> PathBuf {
-	let file_name = source
-		.file_name()
-		.unwrap_or_else(|| std::ffi::OsStr::new("unknown"));
-	let backup_name = format!("reinhardt-fmt-{}.bak", file_name.to_string_lossy());
-	std::env::temp_dir().join(backup_name)
-}
-
-/// Mask sensitive file path in error messages.
-///
-/// Returns a masked version of the path that only shows the filename,
-/// preventing full path disclosure in error output.
-fn mask_path(path: &Path) -> String {
-	path.file_name()
-		.map(|name| format!("<...>/{}", name.to_string_lossy()))
-		.unwrap_or_else(|| "<file>".to_string())
-}
-
-/// Convert a path to a display-safe relative path.
-///
-/// Attempts to strip the current working directory prefix to show a relative path.
-/// Falls back to showing only the filename via `mask_path` if relativization fails.
-fn display_path(path: &Path) -> String {
-	if let Ok(cwd) = std::env::current_dir()
-		&& let Ok(relative) = path.strip_prefix(&cwd)
-	{
-		return relative.display().to_string();
-	}
-	mask_path(path)
-}
-
-/// Sanitize error messages to prevent information leakage.
-///
-/// Removes absolute file system paths and sensitive patterns from error messages
-/// to prevent exposing internal system structure to end users.
-fn sanitize_error(error: &str) -> String {
-	use std::sync::LazyLock;
-
-	// Pattern: absolute paths on Unix-like systems (e.g., /home/user/project/file.rs)
-	static PATH_RE: LazyLock<regex::Regex> =
-		LazyLock::new(|| regex::Regex::new(r"(/[a-zA-Z0-9._-]+){3,}").unwrap());
-
-	// Pattern: database connection URLs
-	static DB_RE: LazyLock<regex::Regex> = LazyLock::new(|| {
-		regex::Regex::new(r"(?i)(postgres|mysql|sqlite|mongodb|redis)://[^\s]+").unwrap()
-	});
-
-	// Pattern: API keys, tokens, secrets in key=value or key:value format
-	static TOKEN_RE: LazyLock<regex::Regex> = LazyLock::new(|| {
-		regex::Regex::new(
-			r"(?i)(api[_\-]?key|token|secret|password|auth)[=:]\s*['\x22]?[a-zA-Z0-9+/=_\-]{8,}",
-		)
-		.unwrap()
-	});
-
-	let sanitized = PATH_RE.replace_all(error, |caps: &regex::Captures| {
-		let matched_path = Path::new(caps.get(0).unwrap().as_str());
-		mask_path(matched_path)
-	});
-
-	let sanitized = DB_RE.replace_all(&sanitized, "[REDACTED_DATABASE_URL]");
-	let sanitized = TOKEN_RE.replace_all(&sanitized, "[REDACTED_CREDENTIAL]");
-
-	sanitized.to_string()
-}
-
-/// Securely clear a HashMap containing sensitive string data.
-///
-/// # Security
-///
-/// Uses zeroize to overwrite string values in memory before dropping,
-/// preventing sensitive data from remaining in memory.
-fn secure_clear_hashmap(map: &mut std::collections::HashMap<PathBuf, String>) {
-	for (_, value) in map.iter_mut() {
-		value.zeroize();
-	}
-	map.clear();
-}
-
-/// Validate a config path argument to prevent path traversal and special file attacks.
-///
-/// # Security
-///
-/// - Rejects paths containing traversal sequences (`..`)
-/// - Verifies the path exists and is a regular file (not a directory, device, etc.)
-/// - Rejects symlinks to prevent symlink-based attacks
-/// - Rejects special device paths (e.g., `/dev/stdin`, `/dev/null`)
-fn validate_config_path(path: &Path) -> Result<(), String> {
-	// Check for path traversal sequences
-	let path_str = path.to_string_lossy();
-	if path_str.contains("..") {
-		return Err(format!(
-			"Config path contains path traversal sequence: {}",
-			mask_path(path)
-		));
-	}
-
-	// Reject special device paths
-	#[cfg(unix)]
-	if path_str.starts_with("/dev/")
-		|| path_str.starts_with("/proc/")
-		|| path_str.starts_with("/sys/")
-	{
-		return Err(format!(
-			"Config path refers to a special device: {}",
-			mask_path(path)
-		));
-	}
-
-	// Use symlink_metadata for a single atomic check (avoids TOCTOU between
-	// exists/is_symlink/is_file calls). symlink_metadata does not follow
-	// symlinks, so a symlink target cannot change between checks.
-	let symlink_meta = std::fs::symlink_metadata(path)
-		.map_err(|e| format!("Config path is not accessible: {} ({})", mask_path(path), e))?;
-
-	// Reject symlinks
-	if symlink_meta.file_type().is_symlink() {
-		return Err(format!(
-			"Config path is a symlink, which is not allowed: {}",
-			mask_path(path)
-		));
-	}
-
-	// Verify it's a regular file (not a directory, device, etc.)
-	if !symlink_meta.is_file() {
-		return Err(format!(
-			"Config path is not a regular file: {}",
-			mask_path(path)
-		));
-	}
-
-	Ok(())
-}
-
-/// Maximum file size for configuration files (10 MB).
-///
-/// Prevents OOM from processing extremely large files.
-const MAX_CONFIG_FILE_SIZE: u64 = 10 * 1024 * 1024;
-
-/// Maximum file size for Rust source files (5 MB).
-///
-/// Prevents OOM from processing extremely large source files.
-const MAX_SOURCE_FILE_SIZE: u64 = 5 * 1024 * 1024;
-
-/// Check file size before reading to prevent OOM.
-///
-/// # Errors
-///
-/// Returns an error if the file exceeds the given maximum size.
-fn check_file_size(path: &Path, max_size: u64) -> Result<(), String> {
-	match std::fs::metadata(path) {
-		Ok(metadata) => {
-			if metadata.len() > max_size {
-				Err(format!(
-					"File {} exceeds maximum allowed size ({} bytes, limit {} bytes)",
-					mask_path(path),
-					metadata.len(),
-					max_size
-				))
-			} else {
-				Ok(())
-			}
-		}
-		Err(e) => Err(format!(
-			"Failed to check file size for {}: {}",
-			mask_path(path),
-			e
-		)),
-	}
-}
-
-/// Acquire an exclusive lock file to prevent concurrent format operations.
-///
-/// # Security
-///
-/// Uses `OpenOptions::create_new(true)` for atomic lock creation (TOCTOU mitigation).
-/// The lock file is automatically removed when the returned guard is dropped.
-///
-/// # Errors
-///
-/// Returns an error if the lock file already exists (another operation is in progress)
-/// or if the file cannot be created.
-fn acquire_format_lock(lock_path: &Path) -> Result<FormatLockGuard, std::io::Error> {
-	use std::fs::OpenOptions;
-
-	// Atomic create-or-fail: prevents TOCTOU race between check and create
-	let file = OpenOptions::new()
-		.write(true)
-		.create_new(true)
-		.open(lock_path)?;
-
-	// Write PID to lock file for debugging
-	use std::io::Write;
-	let mut file = file;
-	let _ = writeln!(file, "{}", std::process::id());
-
-	Ok(FormatLockGuard {
-		path: lock_path.to_path_buf(),
-	})
-}
-
-/// RAII guard that removes the lock file on drop.
-struct FormatLockGuard {
-	path: PathBuf,
-}
-
-impl Drop for FormatLockGuard {
-	fn drop(&mut self) {
-		let _ = std::fs::remove_file(&self.path);
+		Err(reinhardt_commands::CommandError::ExecutionError(format!(
+			"reinhardt-formatter exited with status {status}"
+		)))
 	}
 }
