+++ crates/reinhardt-formatter/Cargo.toml
@@ -1,0 +1,32 @@
+[package]
+name = "reinhardt-formatter"
+version = "0.2.0-rc.2"
+edition.workspace = true
+rust-version.workspace = true
+license.workspace = true
+repository.workspace = true
+authors.workspace = true
+description = "Formatter for Reinhardt page!, form!, and head! DSL macros"
+publish = false
+
+[dependencies]
+clap = { version = "4.5", features = ["derive"] }
+colored = { workspace = true }
+regex = "1.10"
+tree-sitter = "0.26"
+tree-sitter-rust = "0.24.2"
+tree-sitter-reinhardt-page = { workspace = true }
+tree-sitter-reinhardt-form = { workspace = true }
+tree-sitter-reinhardt-head = { workspace = true }
+topiary-core = { workspace = true }
+topiary-tree-sitter-facade = { workspace = true }
+walkdir = "2.5"
+zeroize = "1.8"
+
+[dev-dependencies]
+rstest = { workspace = true }
+tempfile = "3.8"
+
+[[bin]]
+name = "reinhardt-formatter"
+path = "src/main.rs"
